# api/docs/swagger.yaml
definitions:
...
  rest.Languages:
    properties:
      en:
        type: string
      ru:
        type: string
      uz:
        type: string
    type: object
...
  rest.districtParams:
    properties:
      code:
        type: string
      code_gni:
        type: string
      code_province:
        type: string
      code_soato:
        type: string
      is_active:
        type: boolean
      title:
        $ref: '#/definitions/rest.Languages'
    required:
    - code
    - code_gni
    - code_province
    - code_soato
    - is_active
    - title
    type: object
...
  rest.nationalityParams:
    properties:
      code:
        type: string
      state:
        type: boolean
      title:
        $ref: '#/definitions/rest.Languages'
    required:
    - code
    - state
    - title
    type: object
...
  rest.updateDistrictParams:
    properties:
      code_gni:
        type: string
      code_province:
        type: string
      code_soato:
        type: string
      is_active:
        type: boolean
      title:
        $ref: '#/definitions/rest.Languages'
    required:
    - code_gni
    - code_province
    - code_soato
    - is_active
    - title
    type: object
...
  rest.updateNationalityParams:
    properties:
      state:
        type: boolean
      title:
        $ref: '#/definitions/rest.Languages'
    required:
    - state
    - title
    type: object
...
  views.Languages:
    properties:
      en:
        type: string
      ru:
        type: string
      uz:
        type: string
    type: object
...
  views.R:
    properties:
      data: {}
      error_code:
        type: integer
      error_note:
        type: string
      status:
        type: string
    type: object
  views.district:
    properties:
      code:
        type: string
      code_gni:
        type: string
      code_province:
        type: string
      code_soato:
        type: string
      is_active:
        type: boolean
      title:
        $ref: '#/definitions/views.Languages'
    type: object
  views.nationality:
    properties:
      code:
        type: string
      is_active:
        type: boolean
      title:
        $ref: '#/definitions/views.Languages'
    type: object
...
info:
  contact:
    email: sodikjon.akhmedoff@gmail.com
    name: API Support
    url: https://github.com/s-akhmedoff
  description: CRUD API for all references
  license:
    name: Hamkorbank
    url: https://hamkorbank.uz/
  title: Reference Service v1
paths:
...
  /districts:
    get:
      consumes:
      - application/json
      description: API to get all district
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/views.R'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/views.district'
                  type: array
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Read all district
      tags:
      - District
    post:
      consumes:
      - application/json
      description: API to create a district
      parameters:
      - description: create district request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.districtParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Create district
      tags:
      - District
  /districts/{code}:
    delete:
      consumes:
      - application/json
      description: API to deactivate district by code
      parameters:
      - description: district code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Delete district
      tags:
      - District
    get:
      consumes:
      - application/json
      description: API to get a district by code
      parameters:
      - description: district code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/views.R'
            - properties:
                data:
                  $ref: '#/definitions/views.district'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Read one district
      tags:
      - District
    put:
      consumes:
      - application/json
      description: API to update district by code
      parameters:
      - description: district code
        in: path
        name: code
        required: true
        type: string
      - description: update district request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.updateDistrictParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Update district
      tags:
      - District
  /districts/import:
    put:
      consumes:
      - multipart/form-data
      description: API to deactivate district by code
      parameters:
      - description: 'Excel File in format: A:CodeSOATO|B:Code|C:CodeGNI|D:CodeProvince|E:NameRu|F:NameEn|G:NameUz'
        in: formData
        name: excel
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Create district
      tags:
      - District
...
  /nationalities:
    get:
      consumes:
      - application/json
      description: API to get all nationalities
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/views.R'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/views.nationality'
                  type: array
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Read all nationality
      tags:
      - Nationality
    post:
      consumes:
      - application/json
      description: API to create a nationality
      parameters:
      - description: create nationality request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.nationalityParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Create nationality
      tags:
      - Nationality
  /nationalities/{code}:
    delete:
      consumes:
      - application/json
      description: API to deactivate nationality by code
      parameters:
      - description: nationality code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Delete nationality
      tags:
      - Nationality
    get:
      consumes:
      - application/json
      description: API to get a nationality by code
      parameters:
      - description: nationality code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/views.R'
            - properties:
                data:
                  $ref: '#/definitions/views.nationality'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Read one nationality
      tags:
      - Nationality
    put:
      consumes:
      - application/json
      description: API to update nationality by code
      parameters:
      - description: nationality code
        in: path
        name: code
        required: true
        type: string
      - description: update nationality request parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/rest.updateNationalityParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Update nationality
      tags:
      - Nationality
  /nationalities/import:
    put:
      consumes:
      - multipart/form-data
      description: API to deactivate nationality by code
      parameters:
      - description: 'Excel File in format: A:Code|B:NameRU|C:NameEn|D:NameUz'
        in: formData
        name: excel
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/views.R'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/views.R'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/views.R'
      summary: Create nationality
      tags:
      - Nationality
...
schemes:
- http
swagger: "2.0"